:construction:: Cration boîte dialog confirmation ajout et suppression dans dashboard

+++ src/app/features/administrateurs-list/administrateurs-list.component.ts
@@ -1,6 +1,9 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Administrateur } from 'src/app/models/Administrateur';
 import { AdministrateurService } from 'src/app/services/administrateur.service';
+import { DialogSuppressionElementComponent } from '../dialog-suppression-element/dialog-suppression-element.component';
+
 
 @Component({
   selector: 'app-administrateurs-list',
@@ -12,7 +15,8 @@ export class AdministrateursListComponent implements OnInit {
   administrateurs: Administrateur[];
   tableColumns: string[] = ['nom', 'prenom', 'email', 'telephone', 'supprimer'];
 
-  constructor(private administrateurService: AdministrateurService) { }
+  constructor(private administrateurService: AdministrateurService, 
+              public dialog: MatDialog) { }
 
   ngOnInit(): void {
     this.loadData();
@@ -22,11 +26,17 @@ export class AdministrateursListComponent implements OnInit {
     this.administrateurService.getAll().subscribe(res => {
     this.administrateurs = res;
     })
-  }
+  }  
 
   deleteAdmin(id: number) {
-    this.administrateurService.deleteById(id).subscribe();   
-    this.administrateurs = this.administrateurs.filter(item => item.id != id);
+    const dialogRef = this.dialog.open(DialogSuppressionElementComponent);
+
+    dialogRef.afterClosed().subscribe(result => {
+       if(result=== true) {
+        this.administrateurService.deleteById(id).subscribe();   
+        this.administrateurs = this.administrateurs.filter(item => item.id != id);
+       }
+    });
   }
 
 }
